requirements.txt
# NetworkChuck AI Chatbot Requirements
# Enhanced with LangChain Agents for Academic Compliance

# === WEB INTERFACE ===
gradio                           # Interactive web UI (latest version for best compatibility)

# === AI AND LANGUAGE MODELS ===
langchain                        # LLM application framework
langchain-core                   # LangChain core components  
langchain-openai                 # OpenAI integration for LangChain
langchain-community              # Community integrations
langchain-text-splitters         # Text splitting utilities
langsmith                       # LangChain monitoring and debugging (for academic requirement)
openai                          # OpenAI API client (latest for best features)

# === VECTOR DATABASES ===
langchain-pinecone              # Pinecone integration for LangChain
pinecone-client                 # Pinecone vector database client (latest version)

# === TRANSCRIPT EXTRACTION ===
openai-whisper>=20231117        # AI-powered speech-to-text (primary method)
yt-dlp>=2023.12.30             # YouTube downloader (for Whisper audio extraction)
# youtube-transcript-api==0.6.2     # YouTube transcript API (fast fallback when available) ### Doesn't work due to geographic blocking

# === CONFIGURATION AND ENVIRONMENT ===
python-dotenv                   # Environment variable management

# === DATA PROCESSING ===
numpy                          # Numerical computing (latest for compatibility)
pandas                         # Data manipulation and analysis (latest version)
tiktoken                       # OpenAI tokenizer

# === PROGRESS BARS AND UTILITIES ===
tqdm>=4.65.0                   # Progress bars for batch processing
jupyter>=1.0.0                 # Jupyter notebooks for analysis  
ipykernel>=6.25.0              # Jupyter kernel

# === LANGCHAIN AGENTS DEPENDENCIES (ACADEMIC COMPLIANCE) ===
pydantic                       # Data validation for LangChain tools
typing-extensions              # Type hints support for agents

# === OPTIONAL ADDITIONS FOR ENHANCED FUNCTIONALITY ===
# Uncomment these if you want additional features:

# streamlit>=1.28.0              # Alternative web interface to Gradio
# chromadb>=0.4.0                # Alternative vector database to Pinecone

# === DEVELOPMENT AND TESTING ===
pytest                         # Testing framework for validation

# === ElevenLabs Python SDK for Text-to-Speech ===
elevenlabs

# === For audio processing (Speech-to-Text) ===
numpy
wave